global-azure/functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
azure-data-tables==12.0.0
pyyaml
openapi-core
openapi-spec-validator
jsonschema==4.17.3
packaging==23.2
